break up connection operations from auth polling
